+++ src/snrt/mod.rs
@@ -1,5 +1,5 @@
 use std::boxed::Box;
-use std::f64;
+use std::f32;
 
 use snmath::Vector3;
 use snmath::Ray;
@@ -14,13 +14,13 @@ pub struct Camera {
     u: Vector3,
     v: Vector3,
     w: Vector3,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
-    pub fn create_camera(look_from: Vector3, look_at: Vector3, v_up: Vector3, v_fov: f64, aspect_ratio: f64, aperture: f64, focal_dist: f64) -> Camera {
+    pub fn create_camera(look_from: Vector3, look_at: Vector3, v_up: Vector3, v_fov: f32, aspect_ratio: f32, aperture: f32, focal_dist: f32) -> Camera {
         let lens_radius = aperture / 2.0;
-        let theta = v_fov * f64::consts::PI / 180.0;
+        let theta = v_fov * f32::consts::PI / 180.0;
         let half_height = (theta / 2.0).tan();
         let half_width = aspect_ratio * half_height;
         
@@ -38,7 +38,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd = self.lens_radius * Vector3::generate_random_unit_disc();
         let jitter = self.u * rd.x + self.v * rd.y;
         Ray {origin:self.origin+jitter, direction:self.lower_left_corner + self.horizontal * s + self.vertical * t - jitter}
@@ -47,25 +47,25 @@ impl Camera {
 
 
 pub trait Hitable {
-    fn hit_process(&self, r: &Ray, t: f64) -> (bool, Ray, Vector3);
-    fn hit_check(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<f64>;
+    fn hit_process(&self, r: &Ray, t: f32) -> (bool, Ray, Vector3);
+    fn hit_check(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<f32>;
 }
 
 pub struct Sphere {
     pub pos: Vector3,
-    pub radius: f64,
+    pub radius: f32,
     pub material: Box<material::Material>,
 }
 
 impl Hitable for Sphere {
-    fn hit_process(&self, r: &Ray, t: f64) -> (bool, Ray, Vector3) {
+    fn hit_process(&self, r: &Ray, t: f32) -> (bool, Ray, Vector3) {
         let hit_pos = r.point_at_parameter(t);
         let hit_normal = (hit_pos - self.pos).normalize();
 
         self.material.scatter(r, hit_pos, hit_normal)
     }
 
-    fn hit_check(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<f64> {
+    fn hit_check(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<f32> {
         let oc = r.origin - self.pos;
 
         let a = r.direction.dot(&r.direction);
